requirements.txt
# requirements.txt for ha-tools
requests
websockets
PyYAML
